// web/src/form/questions/response/__generated__/ResponseModeFileQuestionResponseQuery.graphql.ts
export type ResponseModeFileQuestionResponseQuery$variables = {
  formInstanceID: string;
  questionID: string;
};
export type ResponseModeFileQuestionResponseQuery$data = {
  readonly questionResponses: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly __typename: "QuestionResponse";
        readonly id: string;
        readonly value: string;
      } | null | undefined;
    } | null | undefined> | null | undefined;
  };
};
export type ResponseModeFileQuestionResponseQuery = {
  response: ResponseModeFileQuestionResponseQuery$data;
  variables: ResponseModeFileQuestionResponseQuery$variables;
};